refactor: split synthetic models and sources
